git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/tags/latest-java6-merge@783829 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/luni/src/test/api/common/org/apache/harmony/luni/tests/java/io/FileTest.java
@@ -60,12 +60,14 @@ public class FileTest extends TestCase {
         // Delete all old temporary files
         File tempDir = new File(System.getProperty("java.io.tmpdir"));
         File[] files = tempDir.listFiles();
-        for (File file : files) {
-            if (file.getName().startsWith("harmony-test")) {
-                if (file.isDirectory()) {
-                    deleteTempFolder(file);
+        if (files != null) {
+            for (File file : files) {
+                if (file.getName().startsWith("harmony-test")) {
+                    if (file.isDirectory()) {
+                        deleteTempFolder(file);
+                    }
+                    file.delete();
                 }
-                file.delete();
             }
         }
     }
